+++ frontend/core/containers/thread/DashboardThread/salon/layout/gauss_blur.ts
